refactor(channels,policy): move discord_reader -> channels/discord/reader, discord_context_policy -> policy/context/discord
Churn-safe slice 2 of the phase-1a lane (neither file nor any importer is
touched by #3239). contextNotFrom authorization is policy, not channel
mechanics, so it lands under policy/context/ per the boundary doctrine.
Pure moves + dotted-import rewrite in the two CLI consumers; test loaders
follow the sys.modules keys the new import names produce.
