+++ src/channels/discord/reader.py
